refacto image upload

+++ src/main/java/fr/polytech/melusine/models/dtos/requests/UserUpdateRequest.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @Builder
 public class UserUpdateRequest {
 
-    private long credit;
+    private double credit;
 
 }
